DAG-371: Legger dokumentasjonssiden som eget steg i søknadsflyten før oppsummering
